IOT sensor libraries
I started working on the IOT implementation and faced to many problems while installing libraries for MPU-6050 and MAX30102. most of the time problem was with the MAX30102 module library. finally solved it and check those two modules. they works perfectly. so decide to save those work for future references.